+++ CHANGELOG.md
@@ -10,11 +10,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added:
 
 - `rcli` alias for `reduct-cli`, [PR-10](https://github.com/reduct-storage/reduct-cli/pull/10)
-- Options for server URL and API token to alias add command, [PR-11](https://github.com/reduct-storage/reduct-cli/pull/11)
+- Options for server URL and API token to alias add
+  command, [PR-11](https://github.com/reduct-storage/reduct-cli/pull/11)
+- `rcl bucket ls` to list buckets, [PR-13](https://github.com/reduct-storage/reduct-cli/pull/13)
 
 ### Changed:
 
 - Improve output of server status command, [PR-12](https://github.com/reduct-storage/reduct-cli/pull/12)
+
 ## [0.1.0] - 2022-10-24
 
 ### Added: